django/requirements.txt
 Django>=1.8,<2.0
 psycopg2
 django-webpack-loader
 djangorestframework
 markdown
 django-filter